requirements.txt
# Core Django
Django==5.1.6
asgiref==3.8.1
sqlparse==0.5.3

# Tailwind
django-tailwind==3.8.0
django-browser-reload==1.18.0

# CORS
django-cors-headers==4.7.0

# Essential utilities
python-dotenv==1.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
python-slugify==8.0.4

# Security
requests==2.32.3
urllib3==2.3.0

gunicorn==23.0.0
python-dateutil==2.9.0
whitenoise==6.9.0
psycopg2-binary==2.9.10

django-allauth==65.8.0
PyJWT==2.10.1
